+++ lib/src/compile.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_compile.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: dcb7cfbedf1e1189808c0056debf6a68bd387dab
+// Checksum: 8e813f2ead6e78899ce820e279983278809a7ea5
 //
 // ignore_for_file: unused_import
 
@@ -25,6 +25,7 @@ import 'importer.dart';
 import 'importer/node.dart';
 import 'io.dart';
 import 'logger.dart';
+import 'logger/terse.dart';
 import 'syntax.dart';
 import 'utils.dart';
 import 'visitor/evaluate.dart';
@@ -44,8 +45,13 @@ CompileResult compile(String path,
     bool useSpaces = true,
     int? indentWidth,
     LineFeed? lineFeed,
+    bool quietDeps = false,
+    bool verbose = false,
     bool sourceMap = false,
     bool charset = true}) {
+  TerseLogger? terseLogger;
+  if (!verbose) logger = terseLogger = TerseLogger(logger ?? Logger.stderr());
+
   // If the syntax is different than the importer would default to, we have to
   // parse the file manually and we can't store it in the cache.
   Stylesheet? stylesheet;
@@ -53,14 +59,15 @@ CompileResult compile(String path,
       (syntax == null || syntax == Syntax.forPath(path))) {
     importCache ??= ImportCache.none(logger: logger);
     stylesheet = importCache.importCanonical(
-        FilesystemImporter('.'), p.toUri(canonicalize(path)), p.toUri(path))!;
+        FilesystemImporter('.'), p.toUri(canonicalize(path)),
+        originalUrl: p.toUri(path))!;
   } else {
     stylesheet = Stylesheet.parse(
         readFile(path), syntax ?? Syntax.forPath(path),
         url: p.toUri(path), logger: logger);
   }
 
-  return _compileStylesheet(
+  var result = _compileStylesheet(
       stylesheet,
       logger,
       importCache,
@@ -71,8 +78,12 @@ CompileResult compile(String path,
       useSpaces,
       indentWidth,
       lineFeed,
+      quietDeps,
       sourceMap,
       charset);
+
+  terseLogger?.summarize(node: nodeImporter != null);
+  return result;
 }
 
 /// Like [compileString] in `lib/sass.dart`, but provides more options to
@@ -93,12 +104,17 @@ CompileResult compileString(String source,
     int? indentWidth,
     LineFeed? lineFeed,
     Object? url,
+    bool quietDeps = false,
+    bool verbose = false,
     bool sourceMap = false,
     bool charset = true}) {
+  TerseLogger? terseLogger;
+  if (!verbose) logger = terseLogger = TerseLogger(logger ?? Logger.stderr());
+
   var stylesheet =
       Stylesheet.parse(source, syntax ?? Syntax.scss, url: url, logger: logger);
 
-  return _compileStylesheet(
+  var result = _compileStylesheet(
       stylesheet,
       logger,
       importCache,
@@ -109,8 +125,12 @@ CompileResult compileString(String source,
       useSpaces,
       indentWidth,
       lineFeed,
+      quietDeps,
       sourceMap,
       charset);
+
+  terseLogger?.summarize(node: nodeImporter != null);
+  return result;
 }
 
 /// Compiles [stylesheet] and returns its result.
@@ -127,6 +147,7 @@ CompileResult _compileStylesheet(
     bool useSpaces,
     int? indentWidth,
     LineFeed? lineFeed,
+    bool quietDeps,
     bool sourceMap,
     bool charset) {
   var evaluateResult = evaluate(stylesheet,
@@ -135,6 +156,7 @@ CompileResult _compileStylesheet(
       importer: importer,
       functions: functions,
       logger: logger,
+      quietDeps: quietDeps,
       sourceMap: sourceMap);
 
   var serializeResult = serialize(evaluateResult.stylesheet,