+++ lib/src/executable/options.dart
@@ -99,6 +99,11 @@ class ExecutableOptions {
       ..addFlag('unicode',
           help: 'Whether to use Unicode characters for messages.')
       ..addFlag('quiet', abbr: 'q', help: "Don't print warnings.")
+      ..addFlag('quiet-deps',
+          help: "Don't print compiler warnings from dependencies.\n"
+              "Stylesheets imported through load paths count as dependencies.")
+      ..addFlag('verbose',
+          help: "Print all deprecation warnings even when they're repetitive.")
       ..addFlag('trace', help: 'Print full Dart stack traces for exceptions.')
       ..addFlag('help',
           abbr: 'h', help: 'Print this usage information.', negatable: false)
@@ -163,8 +168,14 @@ class ExecutableOptions {
       ? _options['unicode'] as bool
       : !term_glyph.ascii;
 
-  /// Whether to silence normal output.
+  /// Whether to silence all warnings.
   bool get quiet => _options['quiet'] as bool;
+
+  /// Whether to silence warnings in dependencies.
+  bool get quietDeps => _options['quiet-deps'] as bool;
+
+  /// Whether to emit all repetitive deprecation warnings.
+  bool get verbose => _options['verbose'] as bool;
 
   /// The logger to use to emit messages from Sass.
   Logger get logger => quiet ? Logger.quiet : Logger.stderr(color: color);